+++ sdk/AppNexusSDK/AppNexusSDK.h
@@ -48,3 +48,7 @@ FOUNDATION_EXPORT const unsigned char AppNexusSDKVersionString[];
 
 #import <AppNexusSDK/ANGDPRSettings.h>
 #import <AppNexusSDK/ANUSPrivacySettings.h>
+#import <AppNexusSDK/ANExternalUserId.h>
+
+
+