+++ sdk/sourcefiles/internal/ANUniversalTagRequestBuilder.m
@@ -231,6 +231,13 @@ optionallyWithAdunitMultiAdRequestManager: (nullable ANMultiAdRequest *)adunitMA
         requestDict[@"user"] = user;
     }
     
+    // Set EUID node, EUID - Third party id solutions
+    //
+    NSArray<NSDictionary<NSString *, NSString *> *>  *externalUserIds  = [self externalUserIds];
+    if (externalUserIds && ANAdvertisingTrackingEnabled()) {
+        requestDict[@"eids"] = externalUserIds;
+    }
+    
     NSDictionary<NSString *, id> *device = [self device];
     if (device) {
         requestDict[@"device"] = device;
@@ -563,15 +570,15 @@ optionallyWithAdunitMultiAdRequestManager: (nullable ANMultiAdRequest *)adunitMA
         userDict[@"language"] = language;
     }
     
-    
     //
-    NSString *externalUid = [self.adFetcherDelegate externalUid];
-    if (externalUid) {
-        userDict[@"external_uid"] = externalUid;
+    NSString *publisherUserId = [[ANSDKSettings sharedInstance] publisherUserId];
+    // Use publisherFirstPartyID if it is present. External Id in ANAdProtocol is deprecated.
+    if (publisherUserId) {
+        userDict[@"external_uid"] = publisherUserId;
+    }else if ([self.adFetcherDelegate externalUid]) {
+        userDict[@"external_uid"] = [self.adFetcherDelegate externalUid];
     }
 
-
-    //
     return [userDict copy];
 }
 
@@ -638,17 +645,8 @@ optionallyWithAdunitMultiAdRequestManager: (nullable ANMultiAdRequest *)adunitMA
     deviceDict[@"connectiontype"] = @(connectionType);
 
 
-    //
-    if (@available(iOS 14, *)) {
-#if __has_include(<AppTrackingTransparency/AppTrackingTransparency.h>)
-        if ([ATTrackingManager trackingAuthorizationStatus] == ATTrackingManagerAuthorizationStatusAuthorized){
-            deviceDict[@"limit_ad_tracking"] = [NSNumber numberWithBool:false];
-        }else if ([ATTrackingManager trackingAuthorizationStatus] == ATTrackingManagerAuthorizationStatusDenied){
-            deviceDict[@"limit_ad_tracking"] = [NSNumber numberWithBool:true];
-        }
-#endif
-    }else{
-        deviceDict[@"limit_ad_tracking"] = [NSNumber numberWithBool:!ANAdvertisingTrackingEnabled()];
+    if(ANAdvertisingTrackingEnabled()){
+        deviceDict[@"limit_ad_tracking"] = [NSNumber numberWithBool:NO];
     }
     
     NSDictionary<NSString *, id> *deviceId = [self deviceId];
@@ -710,7 +708,7 @@ optionallyWithAdunitMultiAdRequestManager: (nullable ANMultiAdRequest *)adunitMA
 
 - (NSDictionary<NSString *, id> *)deviceId
 {
-    if([ANGDPRSettings canAccessDeviceData]){
+    if([ANGDPRSettings canAccessDeviceData] && ANAdvertisingTrackingEnabled()){
         return [self fetchAdvertisingIdentifier];
     }
     
@@ -771,6 +769,50 @@ optionallyWithAdunitMultiAdRequestManager: (nullable ANMultiAdRequest *)adunitMA
 
     //
     return [kvSegmentsArray copy];
+}
+
+
+
+- (NSArray<NSDictionary<NSString *, NSString *> *> *)externalUserIds
+{
+    NSArray<ANExternalUserId *>  *externalUserIdArray  = [ANSDKSettings.sharedInstance externalUserIdArray];
+
+    if ([externalUserIdArray count] <= 0)  { return nil; }
+    //
+    NSMutableArray<NSDictionary<NSString *, NSString *> *>  *transformedeuidArray  = [[NSMutableArray<NSDictionary<NSString *, NSString *> *> alloc] init];
+
+    for (ANExternalUserId *externaluserId in externalUserIdArray)
+    {
+        switch (externaluserId.source) {
+            case ANExternalUserIdSourceLiveRamp:
+                [transformedeuidArray addObject:@{
+                                                 @"source"      : @"liveramp.com",
+                                                 @"id"          : externaluserId.userId
+                                             } ];
+                break;
+            case ANExternalUserIdSourceCriteo:
+                [transformedeuidArray addObject:@{
+                                                 @"source"      : @"criteo.com",
+                                                 @"id"          : externaluserId.userId
+                                             } ];
+                break;
+            case ANExternalUserIdSourceNetId:
+                [transformedeuidArray addObject:@{
+                                                 @"source"      : @"netid.de",
+                                                 @"id"          : externaluserId.userId
+                                             } ];
+                break;
+            case ANExternalUserIdSourceTheTradeDesk:
+                [transformedeuidArray addObject:@{
+                                                 @"source"      : @"adserver.org",
+                                                 @"id"          : externaluserId.userId,
+                                                 @"rti_partner"      : @"TDID"
+                                             } ];
+                break;
+        }
+    }
+    //
+    return [transformedeuidArray copy];
 }